requirements.txt
Django==3.1.7
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
Pillow==7.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
django-heroku==0.3.1
psycopg2==2.8.4
PyJWT==1.7.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
